refactor(daemon): extract command execution into commands/handlers.ts
Move the 24 command-execution methods off the Daemon class into a new
`CommandHandlers` class: the `!stop` / `!cancel` / `!resume` / `!queue` /
`/status` / `/fast` / `/models` / `/effort` / `/permission` execution path,
its target-resolution ladder, and the chat-side session controls that share
its cores (Slack status-bar interaction, Discord select card, Telegram
callback tap, the by-key model / effort / permission / fast / output setters).

The parser stays in commands/commands.ts and the pure select projections in
commands/select-projection.ts; handlers.ts is the stateful half that pairs
with them. It owns no Daemon state — every read and single-point write goes
through the new `CommandHost` interface — and the Daemon keeps thin same-name
private delegates, so the ingress paths, the platform connection callbacks
(via the existing PlatformActionSink wiring) and the tests keep calling these
by name. `applyConfiguredRuntimeSettings` / `restoreConfiguredRuntimeSettings`
stay on the Daemon: they are a runtime-config concern, not a command.

No behavior change. daemon.ts −792/+114.

Verification: `pnpm --filter @agentconnect.md/daemon typecheck` clean;
prettier --check and eslint clean on both files; vitest on
daemon-commands, commands, session-action-audit, model-catalog-daemon,
model-key-session, telegram-threading, attribution, no-response,
daemon-agent-mention-routing, daemon-interrupt-safety, daemon-webchat,
command-chrome, platform-contract, platform-registry-drift,
slack-status-actor, discord-app-commands, relay-persist — all passed.
